Add logging for the POST requests
Logging is done to a file and not the webserver logs because,
error_log() truncates the output if its too long. Logging to file
wont truncate the output.
If output is not json encoded, dump the output to the log file
unmanipulated.
Log file should be specified in config file.